Clean up and fix remaining compile warnings

+++ src/amx/float.c
@@ -284,7 +284,7 @@ static cell AMX_NATIVE_CALL n_floatsin(AMX *amx,const cell *params)
     */
     REAL fA = amx_ctof(params[1]);
     fA = ToRadians(fA, params[2]);
-    fA = sin(fA);
+    fA = (float)sin(fA);
     (void)amx;
     return amx_ftoc(fA);
 }
@@ -299,7 +299,7 @@ static cell AMX_NATIVE_CALL n_floatcos(AMX *amx,const cell *params)
     */
     REAL fA = amx_ctof(params[1]);
     fA = ToRadians(fA, params[2]);
-    fA = cos(fA);
+    fA = (float)cos(fA);
     (void)amx;
     return amx_ftoc(fA);
 }
@@ -314,7 +314,7 @@ static cell AMX_NATIVE_CALL n_floattan(AMX *amx,const cell *params)
     */
     REAL fA = amx_ctof(params[1]);
     fA = ToRadians(fA, params[2]);
-    fA = tan(fA);
+    fA = (float)tan(fA);
     (void)amx;
     return amx_ftoc(fA);
 }
